+++ packages/utilities/projects/utilities/src/js/typography/text/textMedium.ts
@@ -1,6 +1,7 @@
-import { fontBehavior, fontFamily, fontStyle, fontVariant, fontWeight } from '../../font';
+import { fontBehavior, fontFamily, fontHyphenation, fontStyle, fontVariant, fontWeight } from '../../font';
 
 export const textMedium = {
   font: `${fontStyle} ${fontVariant} ${fontWeight.regular} 1.5rem/1.5 ${fontFamily}`,
   ...fontBehavior,
+  ...fontHyphenation,
 };