+++ packages/components-angular/tests/e2e/helpers/puppeteer-helper.ts
@@ -26,10 +26,10 @@ export const goto = async (page: Page, url: string) => {
   await waitForComponentsReady(page);
 };
 
-let consoleMessages: ConsoleMessage[] = [];
+const consoleMessages: ConsoleMessage[] = [];
 
 export const initConsoleObserver = (page: Page): void => {
-  consoleMessages = []; // reset
+  consoleMessages.length = 0; // reset
 
   page.on('console', (msg) => {
     consoleMessages.push(msg);