+++ packages/components/src/components/basic/typography/text/text.tsx
@@ -1,14 +1,12 @@
 import { JSX, Component, Prop, h, Element, Host } from '@stencil/core';
 import {
-  calcLineHeightForElement,
   getHTMLElement,
-  transitionListener,
-  attachSlottedCss,
   getDataThemeDarkAttribute,
+  attachSlottedCss,
   attachComponentCss,
+  setLineHeightOnSizeInherit,
 } from '../../../../utils';
 import type { BreakpointCustomizable, TextAlign, TextColor, TextSize, TextWeight, Theme } from '../../../../types';
-import { isSizeInherit } from './text-utils';
 import { getComponentCss, getSlottedCss } from './text-styles';
 
 @Component({
@@ -60,11 +58,7 @@ export class Text {
   }
 
   public componentDidLoad(): void {
-    if (isSizeInherit(this.size)) {
-      transitionListener(this.textTag, 'font-size', () => {
-        this.textTag.style.lineHeight = `${calcLineHeightForElement(this.textTag)}`;
-      });
-    }
+    setLineHeightOnSizeInherit(this.size, this.textTag);
   }
 
   public render(): JSX.Element {