+++ packages/components/src/components/form/select-wrapper/select-wrapper/select-wrapper-styles.ts
@@ -74,5 +74,5 @@ export const getComponentCss = (
 };
 
 export const getSlottedCss = (host: HTMLElement): string => {
-  return getCss(buildSlottedStyles(host, getBaseSlottedStyles()));
+  return getCss(buildSlottedStyles(host, getBaseSlottedStyles({ withDarkTheme: true })));
 };