+++ packages/components/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 - `Table Head Cell` now has a `multiline` property
 
+#### Changed
+- `Headline` has no `hypens` / `overflow-wrap` style by default
+
 ### [2.11.0-skeletons] - 2022-04-21
 
 ### [2.11.0] - 2022-04-21