+++ packages/components/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+#### Added
+- `Table Head Cell` now has a `multiline` property
+
 #### Changed
 - `Headline` has no `hypens` / `overflow-wrap` style by default